Implement and test login
DB changes: passwords are no more in plain, but they are obfuscated;
the password for each user is "password", no more "ciao123".
Insert fields "name" and "surname" for the staff table.

Change the login method in DAO because the type of user must be
inferred in the back end, not transmitted by the front end. Modify
and run unit tests for this method accordingly.